[Test] Change: Move all test to under __tests__
